requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
awslambdaric==2.0.10
boto3==1.34.41
botocore==1.34.49
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
Mako==1.3.2
mangum==0.17.0
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
s3transfer==0.10.0
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.1
